Template of layout config, version 0.1.1
1. Fixed syntax error of single quotes. There is no escaping for single
quotes, so you cannot use a single quote inside a pair of single quotes.
Fortunately, you can use three single quotes, in this case, to quote texts
with a single quote in it.

2. If you are declaring a bracket-surrounded table (surrounded by "[]")
in an array of tables, add the array name before the table name.
For example, if the array is "[[array]]", your table needs to be declared
like this, "[array.table]".

```toml
[[array]]
element1 = 0
[array.table]
element2 = "0"
element3 = "a"

[[array]]
element4 = 1
[array.table]
element5 = "1"
element6 = "b"
```

In JSON, this will be equivalent to:
```json
{
  "array": [
    {
      "element1": 0,
      "table": { "element2": "0", "element3": "a" }
    },
    {
      "element4": 1,
      "table": { "element5": "1", "element6": "b" }
    }
  ]
}
```

Syntax error about array of tables is now fixed in this version.